// web-client/src/types.d.ts
export interface RecordLabelType {
  id: number;
  name: string;
  country: string | null;
  foundationDate: string | null;
  website: string | null;
  logoUrl: string | null;
}